refactor(bridge): extract message action wrapper probing
